@@ -10,23 +10,26 @@
   },
   "pre-commit": "lint-staged",
   "dependencies": {
+    "@notionhq/client": "^2.2.15",
     "@zeit/react-jsx-parser": "2.0.0",
-    "async-sema": "3.1.0",
-    "github-slugger": "1.2.1",
-    "katex": "0.12.0",
-    "next": "^11.1.2",
-    "prismjs": "1.17.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "uuid": "8.1.0"
+    "async-sema": "3.1.1",
+    "clsx": "^2.1.1",
+    "github-slugger": "2.0.0",
+    "katex": "0.16.11",
+    "next": "^14.2.5",
+    "prismjs": "1.29.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-social-media-embed": "^2.5.13",
+    "uuid": "10.0.0"
   },
   "devDependencies": {
-    "@types/katex": "0.11.0",
-    "@types/node": "14.14.31",
-    "@types/react": "17.0.2",
-    "lint-staged": "10.5.4",
+    "@types/katex": "0.16.7",
+    "@types/node": "20.14.11",
+    "@types/react": "18.3.3",
+    "lint-staged": "15.2.7",
     "pre-commit": "1.2.2",
-    "prettier": "2.2.1",
-    "typescript": "^4.4.4"
+    "prettier": "3.3.3",
+    "typescript": "^5.5.4"
   }
 }
