@@ -1,20 +1,21 @@
+'use client'
 import Link from 'next/link'
 import Head from 'next/head'
 import ExtLink from './ext-link'
-import { useRouter } from 'next/router'
+import { usePathname } from 'next/navigation'
 import styles from '../styles/header.module.css'
 
 const navItems: { label: string; page?: string; link?: string }[] = [
   { label: 'Home', page: '/' },
   { label: 'Blog', page: '/blog' },
   { label: 'Contact', page: '/contact' },
-  { label: 'Source Code', link: 'https://github.com/ijjk/notion-blog' },
+  { label: 'Source Code', link: 'https://github.com/lyuai/notion-blog' },
 ]
 
 const ogImageUrl = 'https://notion-blog.now.sh/og-image.png'
 
 const Header = ({ titlePre = '' }) => {
-  const { pathname } = useRouter()
+  const pathname = usePathname()
 
   return (
     <header className={styles.header}>
@@ -34,10 +35,8 @@ const Header = ({ titlePre = '' }) => {
         {navItems.map(({ label, page, link }) => (
           <li key={label}>
             {page ? (
-              <Link href={page}>
-                <a className={pathname === page ? 'active' : undefined}>
-                  {label}
-                </a>
+              <Link href={page} className={pathname === page ? 'active' : undefined}>
+                {label}
               </Link>
             ) : (
               <ExtLink href={link}>{label}</ExtLink>
