@@ -1,13 +1,13 @@
-import Header from '../components/header'
-import ExtLink from '../components/ext-link'
+import Header from '@/components/header'
+import ExtLink from '@/components/ext-link'
 
-import sharedStyles from '../styles/shared.module.css'
-import contactStyles from '../styles/contact.module.css'
+import sharedStyles from '@/styles/shared.module.css'
+import contactStyles from '@/styles/contact.module.css'
 
-import GitHub from '../components/svgs/github'
-import Twitter from '../components/svgs/twitter'
-import Envelope from '../components/svgs/envelope'
-import LinkedIn from '../components/svgs/linkedin'
+import GitHub from '@/components/svgs/github'
+import Twitter from '@/components/svgs/twitter'
+import Envelope from '@/components/svgs/envelope'
+import LinkedIn from '@/components/svgs/linkedin'
 
 const contacts = [
   {
