@@ -5,7 +5,7 @@ export default function Footer() {
     <>
       <footer>
         <span>Deploy your own!</span>
-        <ExtLink href="https://vercel.com/new/git/external?repository-url=https://github.com/ijjk/notion-blog/tree/main&project-name=notion-blog&repository-name=notion-blog">
+        <ExtLink href="https://vercel.com/new/git/external?repository-url=https://github.com/lyuai/notion-blog/tree/main&project-name=notion-blog&repository-name=notion-blog">
           <img
             src="https://vercel.com/button"
             height={46}
